server/requirements.txt
psycopg2
psycopg2-binary
python-dotenv
fastapi[standard]
uvicorn
email-validator
passlib[bcrypt]
pyjwt